Move Hamiltonian summaries to resource estimation
