+++ qamomile/chemistry/__init__.py
@@ -3,21 +3,17 @@
 from qamomile.chemistry.resource_estimation import (
     ChemistryQPEMethod,
     ChemistryQPEModel,
-    PauliHamiltonianResource,
     estimate_qubitized_chemistry_qpe,
     estimate_qubitized_chemistry_qpe_from_model,
     estimate_single_ancilla_trotter_qpe,
     estimate_single_ancilla_trotter_qpe_from_hamiltonian,
-    summarize_pauli_hamiltonian,
 )
 
 __all__ = [
     "ChemistryQPEMethod",
     "ChemistryQPEModel",
-    "PauliHamiltonianResource",
     "estimate_qubitized_chemistry_qpe",
     "estimate_qubitized_chemistry_qpe_from_model",
     "estimate_single_ancilla_trotter_qpe",
     "estimate_single_ancilla_trotter_qpe_from_hamiltonian",
-    "summarize_pauli_hamiltonian",
 ]
